python_example/pyproject.toml
[tool.poetry]
name = "python_example"
version = "0.1.0"
description = ""
authors = ["deepio"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
pillow = "^6.0"
requests = "^2.21"
lxml = "^4.3"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
python_example = 'python_example:commands.cli'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
